reorganize the structure of the decomposition
