+++ pydec/__init__.py
@@ -6,6 +6,7 @@ from .decomposition import (
     get_decomposition_name,
     get_decomposition_func,
     using_decomposition_func,
+    no_decomposition,
     set_decomposition_args,
     get_decomposition_args,
     using_decomposition_args,
