+++ pydec/decomposition/__init__.py
@@ -1,0 +1,11 @@
+from .states import (
+    set_decomposition_func,
+    get_decomposition_name,
+    get_decomposition_func,
+    using_decomposition_func,
+    no_decomposition,
+    set_decomposition_args,
+    get_decomposition_args,
+    using_decomposition_args,
+)
+from . import algorithms
